calculator code
This is the code I created for my calculator.
This code was written for an assignment on my Level 3 BTEC IT Computing Course (120 Credits).
Unit 16: Procedural Programming.

The base code used is generic. It is the small twists that were individual.
we were taught the basics, then "let loose" to code our calculator. We asked help if needed.